customizations-example: Autosize console window when login
Console window has fixed columns and rows, there is no method to pass
the window columns and rows to console terminal.
So it can not adjust the window size automatically.

This change would make adjust the console terminal column and row
according to the actual terminal window size.

The most important benefit is that we can see the whole output when using
journalctl systemctl and so on to output the debug info if the debug
information is too long.

Autosize-console origin from iot2050setup tool.

Signed-off-by: chao zeng <chao.zeng@siemens.com>