Provide the board configuration tools
No matter iot2050setup or board-configuration is a
tool for configuring peripheral devices of the board.
And the configured data is stored in a json file.

So provide board-conf-tools include the iot2050setup
and board-configuration.They use the same interface
to configure the json file.

Therefore, the format of the json is redefined and the
original iot2050setup and board-configuration  file move to
the board-conf-tool recipes.

Signed-off-by: chao zeng <chao.zeng@siemens.com>